+++ src/core/src/utils/shared/helpers.ts
@@ -1,6 +1,10 @@
 import 'reflect-metadata';
 import { DECORATOR } from '../../constants';
-import { ControllerMeta, MiddlewaresMetadata, RouteMetadata } from '../../types/core/controller';
+import {
+  ControllerMeta,
+  MiddlewaresMetadataItem,
+  RouteMetadata,
+} from '../../types/core/controller';
 
 export const generateUniqueId = () => {
   return `${Date.now()}-${Math.random().toString(36).substring(2, 9)}`;
@@ -29,26 +33,21 @@ export function defineControllerMeta(meta: Partial<ControllerMeta>, target: obje
   Reflect.defineMetadata(DECORATOR.controller, merged, target);
 }
 
-export function reflectMiddlewaresMetadata(target: any, property?: string): MiddlewaresMetadata {
+export function reflectMiddlewaresMetadata(
+  target: any,
+  property?: string,
+): MiddlewaresMetadataItem[] {
   let data;
   if (property) {
-    data = Reflect.getMetadata(DECORATOR.middlewares, target, property) ?? {};
+    data = Reflect.getMetadata(DECORATOR.middlewares, target, property) ?? [];
   } else {
-    data = Reflect.getMetadata(DECORATOR.middlewares, target) ?? {};
+    data = Reflect.getMetadata(DECORATOR.middlewares, target) ?? [];
   }
-
-  ['middlewares', 'errors', 'guards', 'pipes', 'cors', 'sanitizers', 'interceptors'].forEach(
-    prop => {
-      if (!data[prop]?.length) {
-        data[prop] = [];
-      }
-    },
-  );
 
   return data;
 }
 export function reflectRouteMetadata(target: object, property: string): RouteMetadata {
-  const data = Reflect.getMetadata(DECORATOR.route, target, property) ?? {};
+  const data = Reflect.getMetadata(DECORATOR.route, target.constructor, property) ?? {};
 
   ['parameters', 'middlewares'].forEach(prop => {
     if (!data[prop]?.length) {
@@ -59,28 +58,20 @@ export function reflectRouteMetadata(target: object, property: string): RouteMet
   return data;
 }
 
-export function defineMiddlewaresMeta(meta: Partial<MiddlewaresMetadata>, target: object): void;
+export function defineMiddlewaresMeta(meta: MiddlewaresMetadataItem[], target: object): void;
 export function defineMiddlewaresMeta(
-  meta: Partial<MiddlewaresMetadata>,
+  meta: MiddlewaresMetadataItem[],
   target: object,
   property?: string,
 ): void;
 export function defineMiddlewaresMeta(
-  meta: Partial<MiddlewaresMetadata>,
+  meta: MiddlewaresMetadataItem[],
   target: object,
   property?: string,
 ) {
   const existed = reflectMiddlewaresMetadata(target, property);
 
-  const merged = Object.entries(meta).reduce((acc, [key, value]) => {
-    if (Array.isArray(acc[key])) {
-      acc[key] = acc[key].concat(value);
-    } else {
-      acc[key] = value;
-    }
-
-    return acc;
-  }, existed as any);
+  const merged = [...meta, ...(existed ?? [])];
   if (property) {
     Reflect.defineMetadata(DECORATOR.middlewares, merged, target, property);
   } else {
@@ -93,7 +84,7 @@ export function defineRouteMeta(
   property: string,
 ): void;
 export function defineRouteMeta(meta: Partial<RouteMetadata>, target: object, property: string) {
-  const existed = reflectRouteMetadata(target, property);
+  const existed = reflectRouteMetadata(target.constructor, property);
   const merged = Object.entries(meta).reduce((acc, [key, value]) => {
     if (Array.isArray(acc[key])) {
       acc[key] = acc[key].concat(value);
@@ -104,5 +95,5 @@ export function defineRouteMeta(meta: Partial<RouteMetadata>, target: object, pr
     return acc;
   }, existed as any);
 
-  Reflect.defineMetadata(DECORATOR.route, merged, target, property);
+  Reflect.defineMetadata(DECORATOR.route, merged, target.constructor, property);
 }