+++ src/middlewares/src/use.ts
@@ -49,9 +49,9 @@ export function Use(middleware: MiddlewareCB | MiddlewareCB[]) {
   return function (target: any, propertyKey?: string, descriptor?: PropertyDescriptor) {
     const middlewares = Array.isArray(middleware) ? middleware : [middleware];
 
-    const data = { middlewares };
+    const data = middlewares.map(middleware => ({ middleware }));
 
-    if (descriptor) {
+    if (propertyKey) {
       defineMiddlewaresMeta(data, target, propertyKey);
     } else {
       defineMiddlewaresMeta(data, target);